Revert "Refactor App.jsx and modularize components"
