+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 // App.jsx - Production-Ready Tranch Platform
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect, createContext, useContext, useRef, useCallback } from 'react';
 import { 
   BrowserRouter as Router, 
   Routes, 
@@ -28,28 +28,9 @@ import {
 } from '@clerk/clerk-react';
 
 import { AppProvider, NotificationProvider, useApp, useNotifications } from "./contexts";
-import {
-  Modal,
-  ConfirmationDialog,
-  NumberInput,
-  ProtectedRoute,
-  DocumentManager,
-  Navigation,
-  BrokerAIFloating,
-  LoadingSpinner,
-  ErrorMessage,
-  SuccessMessage,
-  InfoMessage,
-  Toast,
-  Tooltip,
-  ProgressBar,
-  Tabs,
-  EmptyState,
-  StatusBadge,
-} from "./components";
+import { Modal, ConfirmationDialog, NumberInput, ProtectedRoute, DocumentManager } from "./components";
 import { useApi, useDebounce } from "./hooks";
-import { Onboarding, Dashboard, Projects } from "./pages";
-import { formatCurrency, formatNumber, formatDate, formatDateTime, formatTime, downloadCSV } from "./utils";
+import { Onboarding } from "./pages";
 
 // ===========================
 // CONFIGURATION
@@ -76,9 +57,630 @@ const STRIPE_PUBLISHABLE_KEY = import.meta.env.VITE_STRIPE_PUBLISHABLE_KEY ||
 const stripePromise = loadStripe(STRIPE_PUBLISHABLE_KEY);
 
 // ===========================
+
+// ===========================
+// CONTEXTS
+// UTILITY FUNCTIONS
+// ===========================
+
+const formatCurrency = (amount) => {
+  if (!amount && amount !== 0) return '-';
+  return new Intl.NumberFormat('en-AU', {
+    style: 'currency',
+    currency: 'AUD',
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 0,
+  }).format(amount);
+};
+
+const formatNumber = (num) => {
+  if (!num && num !== 0) return '-';
+  return new Intl.NumberFormat('en-AU').format(num);
+};
+
+const formatDate = (dateString) => {
+  if (!dateString) return '-';
+  return new Date(dateString).toLocaleDateString('en-AU', {
+    day: 'numeric',
+    month: 'short',
+    year: 'numeric',
+  });
+};
+
+const formatDateTime = (dateString) => {
+  if (!dateString) return '-';
+  return new Date(dateString).toLocaleString('en-AU', {
+    day: 'numeric',
+    month: 'short',
+    year: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit'
+  });
+};
+
+const formatTime = (dateString) => {
+  if (!dateString) return '-';
+  const date = new Date(dateString);
+  const now = new Date();
+  const diffInHours = (now - date) / (1000 * 60 * 60);
+  
+  if (diffInHours < 24) {
+    return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
+  } else if (diffInHours < 24 * 7) {
+    return date.toLocaleDateString([], { weekday: 'short' });
+  } else {
+    return date.toLocaleDateString([], { month: 'short', day: 'numeric' });
+  }
+};
+
+const downloadCSV = (data, filename) => {
+  const csv = convertToCSV(data);
+  const blob = new Blob([csv], { type: 'text/csv' });
+  const url = window.URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = filename;
+  a.click();
+  window.URL.revokeObjectURL(url);
+};
+
+const convertToCSV = (data) => {
+  if (!data || data.length === 0) return '';
+  
+  const headers = Object.keys(data[0]);
+  const csvHeaders = headers.join(',');
+  
+  const csvRows = data.map(row => {
+    return headers.map(header => {
+      const value = row[header];
+      // Escape quotes and wrap in quotes if contains comma
+      const escaped = String(value || '').replace(/"/g, '""');
+      return escaped.includes(',') ? `"${escaped}"` : escaped;
+    }).join(',');
+  });
+  
+  return [csvHeaders, ...csvRows].join('\n');
+};
+
+const validateEmail = (email) => {
+  const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return re.test(email);
+};
+
+const validatePhone = (phone) => {
+  const re = /^(\+61|0)[2-478][\d]{8}$/;
+  return re.test(phone.replace(/\s/g, ''));
+};
+
+const validateABN = (abn) => {
+  const abnRegex = /^[0-9]{11}$/;
+  return abnRegex.test(abn.replace(/\s/g, ''));
+};
+
+// ===========================
+// COMPONENTS
+// ===========================
+
+// Loading Spinner
+export const LoadingSpinner = ({ size = 'default', message }) => (
+  <div className={`loading-spinner ${size}`}>
+    <div className="spinner"></div>
+    {message && <p className="loading-message">{message}</p>}
+  </div>
+);
+
+// Error Message
+const ErrorMessage = ({ message, onClose }) => (
+  <div className="error-message">
+    <span>{message}</span>
+    {onClose && <button onClick={onClose} className="close-btn">×</button>}
+  </div>
+);
+
+// Success Message
+const SuccessMessage = ({ message, onClose }) => (
+  <div className="success-message">
+    <span>{message}</span>
+    {onClose && <button onClick={onClose} className="close-btn">×</button>}
+  </div>
+);
+
+// Info Message
+const InfoMessage = ({ message, onClose }) => (
+  <div className="info-message">
+    <span>{message}</span>
+    {onClose && <button onClick={onClose} className="close-btn">×</button>}
+  </div>
+);
+
+// Toast Notification
+const Toast = ({ message, type = 'info', onClose }) => {
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      onClose();
+    }, 5000);
+    return () => clearTimeout(timer);
+  }, [onClose]);
+
+  return (
+    <div className={`toast toast-${type}`}>
+      <span>{message}</span>
+      <button onClick={onClose} className="toast-close">×</button>
+    </div>
+  );
+};
+
+// Enhanced Tooltip Component
+export const Tooltip = ({ children, content, position = 'top' }) => {
+  return (
+    <div className="tooltip-wrapper">
+      {children}
+      <div className={`tooltip tooltip-${position}`}>
+        {content}
+      </div>
+    </div>
+  );
+};
+
+// Modal
+
+// Progress Bar
+const ProgressBar = ({ value, max = 100, label, showPercentage = true }) => {
+  const percentage = Math.min(100, Math.max(0, (value / max) * 100));
+  
+  return (
+    <div className="progress-bar-container">
+      {label && <div className="progress-label">{label}</div>}
+      <div className="progress-bar">
+        <div className="progress-fill" style={{ width: `${percentage}%` }}>
+          {showPercentage && percentage > 10 && (
+            <span className="progress-text">{Math.round(percentage)}%</span>
+          )}
+        </div>
+      </div>
+      {showPercentage && percentage <= 10 && (
+        <span className="progress-text-outside">{Math.round(percentage)}%</span>
+      )}
+    </div>
+  );
+};
+
+// Tab Component
+const Tabs = ({ tabs, activeTab, onChange }) => (
+  <div className="tabs">
+    {tabs.map((tab) => (
+      <button
+        key={tab.id}
+        className={`tab ${activeTab === tab.id ? 'active' : ''}`}
+        onClick={() => onChange(tab.id)}
+        disabled={tab.disabled}
+      >
+        {tab.label}
+        {tab.badge && <span className="tab-badge">{tab.badge}</span>}
+      </button>
+    ))}
+  </div>
+);
+
+// Empty State
+const EmptyState = ({ icon = '📂', title, message, action }) => (
+  <div className="empty-state">
+    <div className="empty-icon">{icon}</div>
+    <h3>{title}</h3>
+    <p>{message}</p>
+    {action && (
+      <div className="empty-action">
+        {action}
+      </div>
+    )}
+  </div>
+);
+
+// Status Badge
+const StatusBadge = ({ status, type = 'default' }) => {
+  const getStatusClass = () => {
+    switch (status?.toLowerCase()) {
+      case 'active':
+      case 'approved':
+      case 'paid':
+      case 'complete':
+      case 'success':
+        return 'success';
+      case 'pending':
+      case 'processing':
+      case 'draft':
+        return 'warning';
+      case 'declined':
+      case 'failed':
+      case 'unpaid':
+      case 'cancelled':
+        return 'error';
+      default:
+        return 'default';
+    }
+  };
+
+  return (
+    <span className={`status-badge status-${getStatusClass()}`}>
+      {status}
+    </span>
+  );
+};
+
+// ===========================
 // NAVIGATION COMPONENT
 // ===========================
 
+const Navigation = () => {
+  const api = useApi();
+  const { user } = useApp();
+  const { notifications, unreadCount, markAllAsRead } = useNotifications();
+  const { signOut } = useClerk();
+  const navigate = useNavigate();
+  const location = useLocation();
+  const [showMobileMenu, setShowMobileMenu] = useState(false);
+  const [showProfileMenu, setShowProfileMenu] = useState(false);
+  const [showNotifications, setShowNotifications] = useState(false);
+  const profileRef = useRef(null);
+  const notificationRef = useRef(null);
+
+
+  
+   useEffect(() => {
+    const handleClickOutside = (event) => {
+      if (profileRef.current && !profileRef.current.contains(event.target)) {
+        setShowProfileMenu(false);
+      }
+      if (notificationRef.current && !notificationRef.current.contains(event.target)) {
+        setShowNotifications(false);
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+    
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, []);
+
+  const handleLogout = async () => {
+    await signOut();
+    navigate('/');
+  };
+
+  if (!user) return null;
+
+  const isActive = (path) => location.pathname === path;
+
+  const navLinks = [
+    { path: '/dashboard', label: 'Dashboard', roles: ['borrower', 'funder', 'admin'] },
+    { path: '/projects', label: 'Projects', roles: ['funder'] },
+    { path: '/my-projects', label: 'My Projects', roles: ['borrower'] },
+    { path: '/messages', label: 'Messages', roles: ['borrower', 'funder'] },
+    { path: '/portfolio', label: 'Portfolio', roles: ['funder'] },
+    { path: '/admin', label: 'Admin', roles: ['admin'] },
+  ];
+
+  const filteredLinks = navLinks.filter(link => link.roles.includes(user.role));
+
+  return (
+    <nav className="navbar">
+      <div className="nav-container">
+        <Link to="/dashboard" className="nav-logo">
+          <span className="logo-text">Tranch</span>
+        </Link>
+        
+        {/* Desktop Navigation */}
+       <div className="nav-menu desktop-only">
+  {filteredLinks.map(link => (
+    <Link 
+      key={link.path}
+      to={link.path} 
+      className={`nav-link ${isActive(link.path) ? 'active' : ''}`}
+    >
+      {link.label}
+    </Link>
+  ))}
+</div>
+
+
+        
+   {/* Mobile Menu Button - Add this check */}
+<button 
+  className="mobile-menu-btn"
+  onClick={() => setShowMobileMenu(!showMobileMenu)}
+>
+  <span></span>
+  <span></span>
+  <span></span>
+</button>
+        
+        {/* User Section */}
+        <div className="nav-user-section">
+          {/* Notifications */}
+          <div className="notification-area" ref={notificationRef}>
+            <button 
+              className={`notification-bell ${unreadCount > 0 ? 'has-notifications' : ''}`}
+              onClick={() => setShowNotifications(!showNotifications)}
+            >
+              <span className="bell-icon">🔔</span>
+              {unreadCount > 0 && (
+                <span className="notification-count">{unreadCount}</span>
+              )}
+            </button>
+            
+            {showNotifications && (
+              <div className="notification-dropdown">
+                <div className="notification-header">
+                  <h3>Notifications</h3>
+                  <button onClick={() => {
+                    markAllAsRead();
+                    setShowNotifications(false);
+                  }}>
+                    Mark all read
+                  </button>
+                </div>
+                
+                <div className="notification-list">
+                  {notifications.length === 0 ? (
+                    <div className="empty-notifications">
+                      <p>No notifications</p>
+                    </div>
+                  ) : (
+                    notifications.slice(0, 10).map(notification => (
+                      <div 
+                        key={notification.id}
+                        className={`notification-item ${notification.read ? 'read' : 'unread'}`}
+                      >
+                        <div className="notification-content">
+                          <strong>{notification.title}</strong>
+                          <p>{notification.message}</p>
+                          <span className="notification-time">
+                            {formatTime(notification.timestamp)}
+                          </span>
+                        </div>
+                      </div>
+                    ))
+                  )}
+                </div>
+              </div>
+            )}
+          </div>
+          
+          {/* Profile Dropdown */}
+          <div className="profile-dropdown-container" ref={profileRef}>
+            <button 
+              className="profile-dropdown-toggle"
+              onClick={() => setShowProfileMenu(!showProfileMenu)}
+            >
+              <div className="user-avatar">
+                {user.name?.charAt(0)?.toUpperCase() || 'U'}
+              </div>
+            </button>
+            
+            {showProfileMenu && (
+              <div className="profile-dropdown-menu">
+                <div className="dropdown-header">
+                  <div className="dropdown-user-info">
+                    <div className="dropdown-user-name">{user.name}</div>
+                    <div className="dropdown-user-role">{user.role}</div>
+                  </div>
+                </div>
+                
+                <div className="dropdown-divider"></div>
+                
+                <Link 
+                  to="/profile" 
+                  className="dropdown-item"
+                  onClick={() => setShowProfileMenu(false)}
+                >
+                  <span className="dropdown-icon">👤</span>
+                  My Profile
+                </Link>
+                
+                <Link 
+                  to="/settings" 
+                  className="dropdown-item"
+                  onClick={() => setShowProfileMenu(false)}
+                >
+                  <span className="dropdown-icon">⚙</span>
+                  Settings
+                </Link>
+                
+                <div className="dropdown-divider"></div>
+                
+                <button onClick={handleLogout} className="dropdown-item logout">
+                  <span className="dropdown-icon">→</span>
+                  Logout
+                </button>
+              </div>
+            )}
+          </div>
+        </div>
+      </div>
+      
+      {/* Mobile Menu */}
+      {showMobileMenu && (
+        <div className="mobile-menu">
+          {filteredLinks.map(link => (
+            <Link 
+              key={link.path}
+              to={link.path} 
+              className={`mobile-menu-link ${isActive(link.path) ? 'active' : ''}`}
+              onClick={() => setShowMobileMenu(false)}
+            >
+              {link.label}
+            </Link>
+          ))}
+          <div className="mobile-menu-divider"></div>
+          <Link 
+            to="/profile" 
+            className="mobile-menu-link"
+            onClick={() => setShowMobileMenu(false)}
+          >
+            My Profile
+          </Link>
+          <Link 
+            to="/settings" 
+            className="mobile-menu-link"
+            onClick={() => setShowMobileMenu(false)}
+          >
+            Settings
+          </Link>
+          <button onClick={handleLogout} className="mobile-menu-link logout">
+            Logout
+          </button>
+        </div>
+      )}
+    </nav>
+
+    
+  );
+};
+
+// ===========================
+// BROKER AI FLOATING ASSISTANT
+// ===========================
+
+const BrokerAIFloating = () => {
+  const api = useApi();
+  const { user } = useApp();
+  const [isOpen, setIsOpen] = useState(false);
+  const [isMinimized, setIsMinimized] = useState(true);
+  const [messages, setMessages] = useState([]);
+  const [input, setInput] = useState('');
+  const [sessionId, setSessionId] = useState(null);
+  const [loading, setLoading] = useState(false);
+  const messagesEndRef = useRef(null);
+
+  useEffect(() => {
+    if (isOpen && !sessionId) {
+      createSession();
+    }
+  }, [isOpen]);
+
+  useEffect(() => {
+    scrollToBottom();
+  }, [messages]);
+
+  const scrollToBottom = () => {
+    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+  };
+
+  const createSession = async () => {
+    try {
+      const response = await api.createAIChatSession(null, 'Quick Chat');
+      setSessionId(response.session_id);
+    } catch (err) {
+      console.error('Failed to create chat session:', err);
+    }
+  };
+
+  const handleSend = async () => {
+    if (!input.trim() || !sessionId || loading) return;
+
+    const userMessage = input.trim();
+    setInput('');
+    setLoading(true);
+
+    // Add user message
+    setMessages(prev => [...prev, {
+      id: Date.now(),
+      sender: 'user',
+      message: userMessage
+    }]);
+
+    try {
+      const response = await api.sendAIChatMessage(sessionId, userMessage);
+      
+      // Add AI response
+      setMessages(prev => [...prev, {
+        id: response.ai_message_id,
+        sender: 'ai',
+        message: response.ai_response
+      }]);
+    } catch (err) {
+      console.error('Failed to send message:', err);
+      setMessages(prev => [...prev, {
+        id: Date.now(),
+        sender: 'ai',
+        message: 'I apologize, but I encountered an error. Please try again.'
+      }]);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  if (!user || isMinimized) {
+    return (
+      <button 
+        className="broker-ai-floating-button"
+        onClick={() => {
+          setIsMinimized(false);
+          setIsOpen(true);
+        }}
+      >
+        <span className="ai-icon">💬</span>
+      </button>
+    );
+  }
+
+  return (
+    <div className={`broker-ai-floating ${isOpen ? 'open' : ''}`}>
+      <div className="broker-ai-header">
+        <h3>BrokerAI Assistant</h3>
+        <div className="broker-ai-controls">
+          <button onClick={() => setIsOpen(false)}>_</button>
+          <button onClick={() => setIsMinimized(true)}>×</button>
+        </div>
+      </div>
+      
+      <div className="broker-ai-messages">
+        {messages.length === 0 ? (
+          <div className="ai-welcome">
+            <p>Hi! I'm your property finance assistant. How can I help you today?</p>
+          </div>
+        ) : (
+          messages.map(msg => (
+            <div key={msg.id} className={`ai-message ${msg.sender}`}>
+              {msg.sender === 'ai' ? (
+                <ReactMarkdown>{msg.message}</ReactMarkdown>
+              ) : (
+                <p>{msg.message}</p>
+              )}
+            </div>
+          ))
+        )}
+        {loading && (
+          <div className="ai-message ai">
+            <div className="typing-indicator">
+              <span></span>
+              <span></span>
+              <span></span>
+            </div>
+          </div>
+        )}
+        <div ref={messagesEndRef} />
+      </div>
+      
+      <div className="broker-ai-input">
+        <input
+          type="text"
+          value={input}
+          onChange={(e) => setInput(e.target.value)}
+          onKeyPress={(e) => e.key === 'Enter' && handleSend()}
+          placeholder="Ask me anything..."
+          disabled={loading}
+        />
+        <button onClick={handleSend} disabled={loading || !input.trim()}>
+          Send
+        </button>
+      </div>
+    </div>
+  );
+};
+
+// ===========================
 // PROTECTED ROUTE
 // ===========================
 
@@ -178,6 +780,647 @@ const ClerkAuthWrapper = ({ mode }) => {
 // ONBOARDING
 // ===========================
 
+const Dashboard = () => {
+  const api = useApi();
+  const { user, refreshUser } = useApp();
+  const { addNotification } = useNotifications();
+  const [projects, setProjects] = useState([]);
+  const [stats, setStats] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState('');
+  const [showSubscriptionModal, setShowSubscriptionModal] = useState(false);
+
+  useEffect(() => {
+    fetchData();
+  }, [user?.role]);
+
+  const fetchData = async () => {
+    try {
+      const projectData = await api.getProjects();
+      setProjects(projectData);
+      
+      if (user?.role === 'admin') {
+        const statsData = await api.getAdminStats();
+        setStats(statsData);
+      }
+    } catch (err) {
+      setError(err.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleProjectUpdate = async () => {
+    await fetchData();
+  };
+
+  if (loading) return <LoadingSpinner />;
+
+  return (
+    <div className="dashboard">
+      <div className="dashboard-header">
+        <div className="header-content">
+          <div className="header-text">
+            <h1>Dashboard</h1>
+            <p className="dashboard-subtitle">
+              {user.role === 'borrower' && 'Manage your property development projects'}
+              {user.role === 'funder' && 'Discover investment opportunities'}
+              {user.role === 'admin' && 'Platform administration'}
+            </p>
+          </div>
+        </div>
+      </div>
+
+      {error && <ErrorMessage message={error} onClose={() => setError('')} />}
+
+      {user.role === 'funder' && !user.approved && (
+        <div className="warning-message">
+          <h3>Account Pending Approval</h3>
+          <p>Your account is currently under review. You'll be able to access projects once approved by our team.</p>
+        </div>
+      )}
+
+      {user.role === 'funder' && user.approved && user.subscription_status !== 'active' && (
+        <div className="subscription-banner">
+          <div className="banner-content">
+            <h3>Activate Your Subscription</h3>
+            <p>Subscribe to unlock full access to all projects and features</p>
+          </div>
+          <button 
+            onClick={() => setShowSubscriptionModal(true)}
+            className="btn btn-primary"
+          >
+            Subscribe Now - $299/month
+          </button>
+        </div>
+      )}
+
+      {user.role === 'admin' && stats && (
+        <div className="admin-stats">
+          <div className="stat-card">
+            <div className="stat-icon">👥</div>
+            <div className="stat-content">
+              <div className="stat-value">{formatNumber(stats.total_users)}</div>
+              <div className="stat-label">Total Users</div>
+            </div>
+          </div>
+          <div className="stat-card">
+            <div className="stat-icon">📁</div>
+            <div className="stat-content">
+              <div className="stat-value">{formatNumber(stats.total_projects)}</div>
+              <div className="stat-label">Total Projects</div>
+            </div>
+          </div>
+          <div className="stat-card">
+            <div className="stat-icon">✓</div>
+            <div className="stat-content">
+              <div className="stat-value">{formatNumber(stats.active_projects)}</div>
+              <div className="stat-label">Active Projects</div>
+            </div>
+          </div>
+          <div className="stat-card">
+            <div className="stat-icon">💰</div>
+            <div className="stat-content">
+              <div className="stat-value">{formatCurrency(stats.total_revenue || 0)}</div>
+              <div className="stat-label">Total Revenue</div>
+            </div>
+          </div>
+        </div>
+      )}
+
+      <div className="projects-section">
+        <div className="section-header">
+          <h2>
+            {user.role === 'borrower' && 'Your Projects'}
+            {user.role === 'funder' && `Available Projects (${projects.length})`}
+            {user.role === 'admin' && 'All Projects'}
+          </h2>
+          {user.role === 'borrower' && (
+            <Link to="/create-project" className="btn btn-primary">
+              <span>+</span> Create New Project
+            </Link>
+          )}
+        </div>
+
+        {projects.length === 0 ? (
+          <EmptyState 
+            icon="📂"
+            title="No projects found"
+            message={
+              user.role === 'borrower' 
+                ? 'Create your first project to get started.'
+                : 'No projects available at the moment.'
+            }
+            action={
+              user.role === 'borrower' && (
+                <Link to="/create-project" className="btn btn-primary">
+                  Create Project
+                </Link>
+              )
+            }
+          />
+        ) : (
+          <div className="projects-grid">
+            {projects.map((project) => (
+              <ProjectCard 
+                key={project.id} 
+                project={project} 
+                userRole={user.role}
+                onProjectUpdate={handleProjectUpdate}
+              />
+            ))}
+          </div>
+        )}
+      </div>
+
+      <SubscriptionModal 
+        isOpen={showSubscriptionModal}
+        onClose={() => setShowSubscriptionModal(false)}
+        onSuccess={async () => {
+          setShowSubscriptionModal(false);
+          await refreshUser();
+          await fetchData();
+          addNotification({
+            type: 'success',
+            title: 'Subscription Active',
+            message: 'Your subscription is now active. You have full access to all projects.'
+          });
+        }}
+      />
+    </div>
+  );
+};
+
+// ===========================
+// PROJECT CARD COMPONENT
+// ===========================
+
+const ProjectCard = ({ project, userRole, onProjectUpdate }) => {
+  const api = useApi();
+  const { addNotification } = useNotifications();
+  const [showPaymentModal, setShowPaymentModal] = useState(false);
+  const [requesting, setRequesting] = useState(false);
+  const [showMessageInput, setShowMessageInput] = useState(false);
+  const [accessMessage, setAccessMessage] = useState('');
+  const navigate = useNavigate();
+
+  const handleRequestAccess = async () => {
+    setRequesting(true);
+    try {
+      await api.requestAccess(project.id, accessMessage.trim() || null);
+      addNotification({
+        type: 'success',
+        title: 'Access Request Sent',
+        message: 'Your request has been sent to the developer.'
+      });
+      setShowMessageInput(false);
+      setAccessMessage('');
+      if (onProjectUpdate) onProjectUpdate();
+    } catch (err) {
+      addNotification({
+        type: 'error',
+        title: 'Request Failed',
+        message: err.message
+      });
+    } finally {
+      setRequesting(false);
+    }
+  };
+
+  const handlePaymentSuccess = async () => {
+    setShowPaymentModal(false);
+    addNotification({
+      type: 'success',
+      title: 'Payment Successful',
+      message: 'Your project is now published and visible to funders.'
+    });
+    
+    if (onProjectUpdate) {
+      await onProjectUpdate();
+    }
+  };
+
+  const getRiskRatingColor = (rating) => {
+    switch (rating?.toLowerCase()) {
+      case 'low': return 'var(--green-600)';
+      case 'medium': return 'var(--yellow-600)';
+      case 'high': return 'var(--red-600)';
+      default: return 'var(--gray-600)';
+    }
+  };
+
+  return (
+    <div className="project-card enhanced">
+      <PaymentModal 
+        isOpen={showPaymentModal}
+        onClose={() => setShowPaymentModal(false)}
+        project={project}
+        onSuccess={handlePaymentSuccess}
+      />
+    
+      <div className="project-header">
+        <div className="header-content">
+          <h3 className="project-title">{project.title}</h3>
+          <div className="project-badges">
+            <StatusBadge status={project.payment_status === 'paid' ? 'Published' : 'Unpublished'} />
+            {project.documents_complete && (
+              <StatusBadge status="Docs Complete" />
+            )}
+          </div>
+        </div>
+        <div className="project-meta">
+          <span className="project-location">📍 {project.suburb}</span>
+          <span className="project-date">📅 {formatDate(project.created_at)}</span>
+        </div>
+      </div>
+
+      <div className="project-financial">
+        <div className="financial-item primary">
+          <label>Loan Amount</label>
+          <span className="value">{formatCurrency(project.loan_amount)}</span>
+        </div>
+        {project.interest_rate && (
+          <div className="financial-item">
+            <label>Interest Rate</label>
+            <span className="value">{project.interest_rate}%</span>
+          </div>
+        )}
+        {project.loan_term && (
+          <div className="financial-item">
+            <label>Loan Term</label>
+            <span className="value">{project.loan_term} months</span>
+          </div>
+        )}
+        {project.lvr && (
+          <div className="financial-item">
+            <label>LVR</label>
+            <span className="value">{project.lvr.toFixed(1)}%</span>
+          </div>
+        )}
+      </div>
+
+      <div className="project-details">
+        <div className="detail-grid">
+          <div className="detail-item">
+            <span className="label">Property Type</span>
+            <span className="value">{project.property_type || 'Not specified'}</span>
+          </div>
+          <div className="detail-item">
+            <span className="label">Development Stage</span>
+            <span className="value">{project.development_stage || 'Planning'}</span>
+          </div>
+          {project.total_project_cost && (
+            <div className="detail-item">
+              <span className="label">Total Project Cost</span>
+              <span className="value">{formatCurrency(project.total_project_cost)}</span>
+            </div>
+          )}
+          {project.expected_profit && (
+            <div className="detail-item">
+              <span className="label">Expected Profit</span>
+              <span className="value">{formatCurrency(project.expected_profit)}</span>
+            </div>
+          )}
+        </div>
+      </div>
+
+      {(project.market_risk_rating || project.construction_risk_rating || project.location_risk_rating) && (
+        <div className="project-risks">
+          <h4>Risk Assessment</h4>
+          <div className="risk-grid">
+            {project.market_risk_rating && (
+              <div className="risk-item">
+                <span className="risk-label">Market Risk</span>
+                <span className="risk-value" style={{ color: getRiskRatingColor(project.market_risk_rating) }}>
+                  {project.market_risk_rating.toUpperCase()}
+                </span>
+              </div>
+            )}
+            {project.construction_risk_rating && (
+              <div className="risk-item">
+                <span className="risk-label">Construction Risk</span>
+                <span className="risk-value" style={{ color: getRiskRatingColor(project.construction_risk_rating) }}>
+                  {project.construction_risk_rating.toUpperCase()}
+                </span>
+              </div>
+            )}
+            {project.location_risk_rating && (
+              <div className="risk-item">
+                <span className="risk-label">Location Risk</span>
+                <span className="risk-value" style={{ color: getRiskRatingColor(project.location_risk_rating) }}>
+                  {project.location_risk_rating.toUpperCase()}
+                </span>
+              </div>
+            )}
+          </div>
+        </div>
+      )}
+
+      {project.description && (
+        <div className="project-description">
+          <p>{project.description}</p>
+        </div>
+      )}
+
+      <div className="project-actions">
+        {userRole === 'borrower' && (
+          <>
+            {project.payment_status === 'unpaid' && (
+              <button 
+                onClick={() => setShowPaymentModal(true)}
+                disabled={!project.documents_complete}
+                className="btn btn-primary"
+                title={!project.documents_complete ? 'Upload all required documents first' : ''}
+              >
+                Pay to Publish ($499)
+              </button>
+            )}
+            <button 
+              onClick={() => navigate(`/project/${project.id}`)}
+              className="btn btn-outline"
+            >
+              View Details
+            </button>
+          </>
+        )}
+
+        {userRole === 'funder' && project.payment_status === 'paid' && (
+          <>
+            {project.access_status !== 'approved' && !showMessageInput && (
+              <button 
+                onClick={() => setShowMessageInput(true)}
+                disabled={project.access_status === 'pending'}
+                className="btn btn-primary"
+              >
+                {project.access_status === 'pending' ? '⏳ Request Pending' : '🔓 Request Full Access'}
+              </button>
+            )}
+            {project.access_status === 'approved' && (
+              <button 
+                onClick={() => navigate(`/project/${project.id}`)}
+                className="btn btn-primary"
+              >
+                View Full Details
+              </button>
+            )}
+            {project.access_status === 'approved' && !project.deal_id && (
+  <button 
+    onClick={async () => {
+      try {
+        const response = await api.createDeal(project.id, project.access_request_id);
+        addNotification({
+          type: 'success',
+          title: 'Deal Room Created',
+          message: 'Successfully created deal room'
+        });
+        navigate(`/project/${project.id}/deal/${response.deal_id}`);
+      } catch (err) {
+        console.error('Deal creation error:', err);
+        addNotification({
+          type: 'error',
+          title: 'Failed to create deal room',
+          message: err.message || 'Could not create deal room'
+        });
+      }
+    }}
+    className="btn btn-primary"
+  >
+    Engage
+  </button>
+)}
+
+{project.deal_id && (
+  <button 
+    onClick={() => navigate(`/project/${project.id}/deal/${project.deal_id}`)}
+    className="btn btn-primary"
+  >
+    Deal Room
+  </button>
+)}
+          </>
+        )}
+
+        {userRole === 'admin' && (
+          <button 
+            onClick={() => navigate(`/project/${project.id}`)}
+            className="btn btn-outline"
+          >
+            Admin View
+          </button>
+        )}
+
+        {userRole === 'funder' && showMessageInput && (
+          <div className="access-request-form">
+            <div className="message-input-container">
+              <label>Message to developer (optional):</label>
+              <textarea
+                value={accessMessage}
+                onChange={(e) => setAccessMessage(e.target.value)}
+                placeholder="Introduce yourself and explain your interest in this project..."
+                className="message-textarea"
+                rows="3"
+                maxLength="500"
+              />
+              <div className="character-count">{accessMessage.length}/500</div>
+            </div>
+            <div className="message-actions">
+              <button 
+                onClick={() => {
+                  setShowMessageInput(false);
+                  setAccessMessage('');
+                }}
+                className="btn btn-sm btn-outline"
+              >
+                Cancel
+              </button>
+              <button 
+                onClick={handleRequestAccess}
+                disabled={requesting}
+                className="btn btn-sm btn-primary"
+              >
+                {requesting ? 'Sending...' : 'Send Request'}
+              </button>
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+};
+
+// ===========================
+// PROJECTS PAGE (FOR FUNDERS)
+// ===========================
+
+const ProjectsPage = () => {
+  const api = useApi();
+  const { user } = useApp();
+  const [projects, setProjects] = useState([]);
+  const [filteredProjects, setFilteredProjects] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState('');
+  const [filters, setFilters] = useState({
+    propertyType: '',
+    minLoan: '',
+    maxLoan: '',
+    suburb: '',
+    developmentStage: '',
+    sortBy: 'created_at'
+  });
+
+  const debouncedFilters = useDebounce(filters, 300);
+
+  useEffect(() => {
+    fetchProjects();
+  }, []);
+
+  useEffect(() => {
+    applyFilters();
+  }, [projects, debouncedFilters]);
+
+  const fetchProjects = async () => {
+    try {
+      const data = await api.getProjects();
+      setProjects(data);
+    } catch (err) {
+      setError(err.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const applyFilters = () => {
+    let filtered = [...projects];
+
+    if (filters.propertyType) {
+      filtered = filtered.filter(p => p.property_type === filters.propertyType);
+    }
+    
+    if (filters.minLoan) {
+      filtered = filtered.filter(p => p.loan_amount >= parseInt(filters.minLoan));
+    }
+    
+    if (filters.maxLoan) {
+      filtered = filtered.filter(p => p.loan_amount <= parseInt(filters.maxLoan));
+    }
+    
+    if (filters.suburb) {
+      filtered = filtered.filter(p => 
+        p.suburb?.toLowerCase().includes(filters.suburb.toLowerCase())
+      );
+    }
+    
+    if (filters.developmentStage) {
+      filtered = filtered.filter(p => p.development_stage === filters.developmentStage);
+    }
+
+    // Sorting
+    filtered.sort((a, b) => {
+      switch (filters.sortBy) {
+        case 'loan_amount_asc':
+          return a.loan_amount - b.loan_amount;
+        case 'loan_amount_desc':
+          return b.loan_amount - a.loan_amount;
+        case 'created_at':
+        default:
+          return new Date(b.created_at) - new Date(a.created_at);
+      }
+    });
+
+    setFilteredProjects(filtered);
+  };
+
+  const clearFilters = () => {
+    setFilters({
+      propertyType: '',
+      minLoan: '',
+      maxLoan: '',
+      suburb: '',
+      developmentStage: '',
+      sortBy: 'created_at'
+    });
+  };
+
+  if (loading) return <LoadingSpinner />;
+
+  if (!user.approved) {
+    return (
+      <div className="projects-page">
+        <div className="warning-message">
+          <h3>Account Pending Approval</h3>
+          <p>Your account is currently under review. You'll be able to access projects once approved.</p>
+        </div>
+      </div>
+    );
+  }
+
+  if (user.subscription_status !== 'active') {
+    return (
+      <div className="projects-page">
+        <div className="subscription-required">
+          <h2>Subscription Required</h2>
+          <p>You need an active subscription to browse projects.</p>
+          <Link to="/dashboard" className="btn btn-primary">
+            Subscribe Now
+          </Link>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="projects-page">
+      <div className="page-header">
+        <h1>Investment Opportunities</h1>
+        <p>Browse and filter active property development projects</p>
+      </div>
+
+      {error && <ErrorMessage message={error} onClose={() => setError('')} />}
+
+      <div className="filters-section">
+        <h3>Filter Projects</h3>
+        <div className="filters-grid">
+          <div className="filter-group">
+            <label>Property Type</label>
+            <select
+              value={filters.propertyType}
+              onChange={(e) => setFilters({ ...filters, propertyType: e.target.value })}
+              className="form-select"
+            >
+              <option value="">All Types</option>
+              <option value="Residential">Residential</option>
+              <option value="Commercial">Commercial</option>
+              <option value="Mixed Use">Mixed Use</option>
+              <option value="Industrial">Industrial</option>
+              <option value="Retail">Retail</option>
+            </select>
+          </div>
+
+          <div className="filter-group">
+            <label>Min Loan Amount</label>
+            <NumberInput
+              value={filters.minLoan}
+              onChange={(value) => setFilters({ ...filters, minLoan: value })}
+              placeholder="Min amount"
+              prefix="$"
+            />
+          </div>
+
+          <div className="filter-group">
+            <label>Max Loan Amount</label>
+            <NumberInput
+              value={filters.maxLoan}
+              onChange={(value) => setFilters({ ...filters, maxLoan: value })}
+              placeholder="Max amount"
+              prefix="$"
+            />
+          </div>
+
+          <div className="filter-group">
+            <label>Suburb</label>
+            <input
+              type="text"
+              value={filters.suburb}
               onChange={(e) => setFilters({ ...filters, suburb: e.target.value })}
               className="form-input"
               placeholder="Search suburb"
@@ -197,6 +1440,86 @@ const ClerkAuthWrapper = ({ mode }) => {
               <option value="Construction">Construction</option>
               <option value="Near Completion">Near Completion</option>
             </select>
+          </div>
+
+          <div className="filter-group">
+            <label>Sort By</label>
+            <select
+              value={filters.sortBy}
+              onChange={(e) => setFilters({ ...filters, sortBy: e.target.value })}
+              className="form-select"
+            >
+              <option value="created_at">Newest First</option>
+              <option value="loan_amount_desc">Loan Amount (High to Low)</option>
+              <option value="loan_amount_asc">Loan Amount (Low to High)</option>
+            </select>
+          </div>
+
+          <button onClick={clearFilters} className="btn btn-outline">
+            Clear Filters
+          </button>
+        </div>
+
+        <div className="filter-summary">
+          Showing {filteredProjects.length} of {projects.length} projects
+        </div>
+      </div>
+
+      {filteredProjects.length === 0 ? (
+        <EmptyState 
+          icon="🔍"
+          title="No projects match your criteria"
+          message="Try adjusting your filters to see more projects"
+          action={
+            <button onClick={clearFilters} className="btn btn-primary">
+              Clear Filters
+            </button>
+          }
+        />
+      ) : (
+        <div className="projects-grid">
+          {filteredProjects.map((project) => (
+            <ProjectCard 
+              key={project.id} 
+              project={project} 
+              userRole={user.role}
+              onProjectUpdate={fetchProjects}
+            />
+          ))}
+        </div>
+      )}
+    </div>
+  );
+};
+
+// ===========================
+// CREATE PROJECT WIZARD
+// ===========================
+
+const CreateProject = () => {
+  const api = useApi();
+  const navigate = useNavigate();
+  const { addNotification } = useNotifications();
+  const [currentStep, setCurrentStep] = useState(1);
+  const [saving, setSaving] = useState(false);
+  const [formData, setFormData] = useState({
+    // Basic Info
+    title: '',
+    description: '',
+    location: '',
+    suburb: '',
+    property_type: 'Residential',
+    development_stage: 'Planning',
+    
+    // Financial Details
+    loan_amount: '',
+    interest_rate: '',
+    loan_term: '',
+    total_project_cost: '',
+    equity_contribution: '',
+    land_value: '',
+    construction_cost: '',
+    expected_gdc: '',
     expected_profit: '',
     
     // Project Details
@@ -6139,7 +7462,7 @@ const AppLayout = () => {
           {/* Funder Routes */}
           <Route path="/projects" element={
             <ProtectedRoute roles={['funder']}>
-              <Projects />
+              <ProjectsPage />
             </ProtectedRoute>
           } />
           <Route path="/portfolio" element={
