+++ frontend/src/components/Navbar.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Navbar = () => (
+  <nav role="navigation" className="navbar">
+    <h1>Tranch</h1>
+  </nav>
+);
+
+export default Navbar;
