--- frontend/src/components/common/SuccessMessage.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const SuccessMessage = ({ message, onClose }) => (
-  <div className="success-message">
-    <span>{message}</span>
-    {onClose && (
-      <button onClick={onClose} className="close-btn">
-        &times;
-      </button>
-    )}
-  </div>
-);
-
-export default SuccessMessage;
