--- frontend/src/components/common/Toast.jsx
@@ -1,19 +1,0 @@
-import React, { useEffect } from 'react';
-
-const Toast = ({ message, type = 'info', onClose }) => {
-  useEffect(() => {
-    const timer = setTimeout(onClose, 5000);
-    return () => clearTimeout(timer);
-  }, [onClose]);
-
-  return (
-    <div className={`toast toast-${type}`}>
-      <span>{message}</span>
-      <button onClick={onClose} className="toast-close">
-        &times;
-      </button>
-    </div>
-  );
-};
-
-export default Toast;
