--- frontend/src/components/common/index.js
@@ -1,10 +1,0 @@
-export { default as LoadingSpinner } from './LoadingSpinner';
-export { default as ErrorMessage } from './ErrorMessage';
-export { default as SuccessMessage } from './SuccessMessage';
-export { default as InfoMessage } from './InfoMessage';
-export { default as Toast } from './Toast';
-export { default as Tooltip } from './Tooltip';
-export { default as ProgressBar } from './ProgressBar';
-export { default as Tabs } from './Tabs';
-export { default as EmptyState } from './EmptyState';
-export { default as StatusBadge } from './StatusBadge';
