+++ frontend/src/components/index.js
@@ -2,12 +2,7 @@ export * from './DocumentManager';
 export * from './Modal';
 export * from './NumberInput';
 export * from './ProtectedRoute';
-export * from './common';
-export { default as PaymentModal } from './PaymentModal';
-export { default as SubscriptionModal } from './SubscriptionModal';
-export { default as ProjectCard } from './ProjectCard';
 
 export * from "./DocumentRequestModal";
 export * from "./FileIcon";
-export { default as Navigation } from './Navigation';
-export { default as BrokerAIFloating } from './BrokerAIFloating';
+export { default as Navbar } from "./Navbar";
