+++ frontend/src/pages/index.js
@@ -1,4 +1,2 @@
 export { default as Onboarding } from './Onboarding';
-export { default as Dashboard } from './Dashboard';
-export { default as Projects } from './Projects';
