--- frontend/src/utils/csv.js
@@ -1,28 +1,0 @@
-export const downloadCSV = (data, filename) => {
-  const csv = convertToCSV(data);
-  const blob = new Blob([csv], { type: 'text/csv' });
-  const url = window.URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = filename;
-  a.click();
-  window.URL.revokeObjectURL(url);
-};
-
-export const convertToCSV = (data) => {
-  if (!data || data.length === 0) return '';
-
-  const headers = Object.keys(data[0]);
-  const csvHeaders = headers.join(',');
-
-  const csvRows = data.map(row => {
-    return headers.map(header => {
-      const value = row[header];
-      // Escape quotes and wrap in quotes if contains comma
-      const escaped = String(value || '').replace(/"/g, '""');
-      return escaped.includes(',') ? `"${escaped}"` : escaped;
-    }).join(',');
-  });
-
-  return [csvHeaders, ...csvRows].join('\n');
-};
