+++ frontend/src/utils/index.js
@@ -1,3 +1,2 @@
 export * from './formatters';
 export * from './validators';
-export * from './csv';
